Add new 'Demo' project to solution
This update includes the creation of new components for the 'Demo' project, which have been added to the main solution. It now contains relevant source and header files, as well as its respective Visual Studio project configuration files. The 'DemoApp' class, included in these updates, extends the existing 'GameApp' class.