Add virtual OnUpdate and OnRender methods to GameApp
The OnUpdate and OnRender methods have been added to the GameApp class and called in the Update and Render methods respectively. These new methods are currently empty but are set up as a framework for future updates to game content, fulfilling a need to improve the structure of the code.