Refactor Logger project's structure and add new files
This commit involves a major rework on the Logger project's structure. It includes the deletion of the Logger.cpp file and extensive additions to Logger.vcxproj. The modifications include adjustments to the compile options, newly listing various .cpp and .h files for the Logger and ConsoleLogger components. Furthermore, it altered the setting of precompiled headers in the project configuration from 'NotUsing' to 'Use'. The changes are aimed at improving the organization, maintenance, and readability of the project.