Add minimum log level and loggable flag to logger
In this commit, enhancements have been made to the logger capabilities of the application. We've added the ability to set a minimum log level for `ConsoleLoggerUnicode` and `ConsoleLoggerMultibyte`. This introduces a filter that prevents the logging of entries below the specified level. Additionally, a `SetLogable` method has been implemented to allow control over whether a specific log level can be logged or not. The ILogger interface has also been refactored for better extendibility and efficiency.